// front/enums/events/game.ts
export enum GameEvents {
  Create = "game:create",
  Update = "game:update",
  Destroy = "game:destroy",
  Fire = "game:fire",
  Earn = "game:earn",
  Move = "game:move",
  Attack = "game:attack",
  Over = "game:over",
  Buy = "game:buy",
  UpgradeDefense = "game:upgrade:defense",
  UpgradeKing = "game:upgrade:king",
}